projetos/polymarket-agent/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
web3>=6.0.0
eth-account>=0.10.0
py-clob-client>=0.18.0
anthropic>=0.40.0
requests>=2.31.0
